+++ src/main/java/com/smartx/tower/model/NfsInode.java
@@ -72,7 +72,7 @@ public class NfsInode {
 
   public static final String SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE = "unique_logical_size";
   @SerializedName(SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE)
-  private Long uniqueLogicalSize;
+  private Double uniqueLogicalSize;
 
   public static final String SERIALIZED_NAME_UNIQUE_SIZE = "unique_size";
   @SerializedName(SERIALIZED_NAME_UNIQUE_SIZE)
@@ -365,7 +365,7 @@ public class NfsInode {
   }
 
 
-  public NfsInode uniqueLogicalSize(Long uniqueLogicalSize) {
+  public NfsInode uniqueLogicalSize(Double uniqueLogicalSize) {
     
     this.uniqueLogicalSize = uniqueLogicalSize;
     return this;
@@ -378,12 +378,12 @@ public class NfsInode {
   @javax.annotation.Nullable
   @ApiModelProperty(value = "")
 
-  public Long getUniqueLogicalSize() {
+  public Double getUniqueLogicalSize() {
     return uniqueLogicalSize;
   }
 
 
-  public void setUniqueLogicalSize(Long uniqueLogicalSize) {
+  public void setUniqueLogicalSize(Double uniqueLogicalSize) {
     this.uniqueLogicalSize = uniqueLogicalSize;
   }
